Refactor neighbor data structures to use Uint16Array for improved memory efficiency, accommodating entity IDs and neighbor counts up to 65535. Update related components across Grid, GameObject, Scene, and workers to ensure consistency and optimize performance.

+++ src/core/utils.js
@@ -1293,7 +1293,7 @@ export function query(componentClasses) {
 
   // Not available in main thread context
   console.warn('[query] Query system only available in worker context');
-  return new Int32Array(0);
+  return new Uint16Array(0);
 }
 
 // ============================================================================
